Extract methods out into match_util
